refactor(analysis): co-locate point enclosure kernels
